# Conflicts:
#	app/src/main/java/aayaamlabs/com/gitdemo/DemoClass.java

+++ app/src/main/java/aayaamlabs/com/gitdemo/DemoClass.java
@@ -11,7 +11,7 @@ public class DemoClass {
     }
 
     public String getGetName() {
-        int a=2;
+        float a=2f;
         return getName;
     }
